AlteraJogo_2022li1g007

Description

Módulo para a realização da Tarefa 6 do projeto de LI1 em 2022/23.

Documentation

estadoInicial :: Jogo

jogoRandom :: Jogo -> Int -> Jogo

mapaF :: StdGen -> Mapa -> Mapa

estadoInicialM :: JogoM

jogoRandomM :: JogoM -> Int -> JogoM

estadoGlossInicial :: Imagens -> Fundo -> Skins -> EstadoGloss

reageEvento :: Event -> EstadoGloss -> EstadoGloss

reageMenu :: Event -> EstadoGloss -> EstadoGloss

escolheModo :: Event -> EstadoGloss -> EstadoGloss

escolhePersonagem :: Event -> EstadoGloss -> EstadoGloss

perdeuJogo :: Event -> EstadoGloss -> EstadoGloss

modoJogo :: Event -> EstadoGloss -> EstadoGloss

pausaJogo :: Event -> EstadoGloss -> EstadoGloss

reageTempo :: Float -> EstadoGloss -> EstadoGloss

calculaPontuacao :: Int -> Jogador -> Pontuacao

A função calculaPontuacao determina quantos pontos o Jogador realizou no final do Jogo, ou seja, verifica quantos terrenos este avançou. De modo, que como o jogador começa na posição 10, a posição inicial do jogador no final do jogo , com as linhas adicionadas desdo início do jogo com a deslizaJogo, será 10 + número de linhas extra.

Para saber onde o jogador perdeu, apenas subtraimos a posição inicial com a posição final do Jogador (10+nLinhas - yfinal).

A função poderia ser definida da seguinte forma:

calculaPontuacao :: Int -> Jogador -> Pontuacao
calculaPontuacao nLinhasExtra (Jogador (_,yf)) = 10 + nLinhasExtra - yf

Exemplos de utilização:

>>> calculaPontuacao 4 (Jogador (3,6))
8

>>> calculaPontuacao 17 (Jogador (3,0))
27