LI12223

Description

Tipos de dados e funções auxiliares para a realização do projeto de LI1 em 2022/23.

Tipos de dados

Básicos

type Coordenadas = (Int, Int)

Par de coordenadas de uma posição no Mapa.

type Largura = Int

Comprimento de um Mapa.

type Velocidade = Int

Velocidade que irá afetar a movimentação dos Obstaculos de um Mapa.

Mapas

data Mapa

O Mapa que constituí o Jogo.

Constructors

Mapa Largura [(Terreno, [Obstaculo])]

data Terreno

Cada linha de um Mapa é um Terreno em particular contendo Obstaculos.

As linhas do tipo Rio ou Estrada têm a propriedade Velocidade que indicam a velocidade de deslocamento dos obstáculos. Podendo esta ser negativa, indicando assim a sua direção.

Constructors

Rio Velocidade

Estrada Velocidade

Relva

data Obstaculo

Um Obstáculo numa linha de um Mapa.

Constructors

Nenhum
a ausência de obstáculos em Rio

Tronco
os troncos deslizam apenas em Rio

Carro
os carros movimentam-se apenas em Estrada

Arvore
as árvores são um obstáculo fixo que não se move e apenas são possíveis em Relva

Jogo

data Jogo

Definição base de um jogo.

Constructors

Jogo

Fields

Jogador

o personagem do jogo

Mapa

o mapa em que se está a jogar

newtype Jogador

O Jogador define o personagem controlado no Jogo.

Constructors

Jogador Coordenadas

data Direcao

Direção de uma Jogada feita por um Jogador no Mapa.

Constructors

Cima

Baixo

Esquerda

Direita

data Jogada

As acções que podem ser tomadas pelo Jogador em cada estado do Jogo.

Constructors

Parado
tipo que define a ausência de uma acção do Jogador

Move Direcao
um movimento do jogador numa determinada Direcao

Jogo MultiPlayer

data JogoM

Tipos de dados extras para a realização do modo MultiPlayer.

Definição base de um jogo multiplayer.

Constructors

JogoM

Fields

Jogador

Uma personagem do jogo multiplayer

Jogador

Outra personagem do jogo multiplayer

Mapa

o mapa em que se está a jogar

Gloss

type Imagem = [(MapaString, (Picture, (Float, Float)))]

Tipos de dados auxiliares para a realização da Interface gráfica do Jogo

Imagem associada a cada Terreno e a cada Obstaculo.

type Fundo = [(String, Picture)]

Imagens do Menu e Pontuação.

type Imagens = [Imagem]

Imagens associadas a cada Terreno e a cada Obstaculo.

type MapaString = String

MapaString é a String associada a cada Terreno e Obstaculo.

type Skins = [(Jogador, String, Picture)]

Lista de imagens que o Jogador pode tomar.

type Pontuacao = Int

Pontuação do Jogador.

type EstadoGloss = (Menu, Jogo, JogoM, Imagens, Fundo, Float, Skins, [Pontuacao])

EstadoGloss é o Estado em que o Jogo se encontra no Gloss.

data Opcao

Opcao são as opções que aparecem no Menu Principal.

Constructors

Jogar

Sair

Instrucoes

data Pausa

Pausa são as opções que aparecem quando o Jogador pausa o Jogo.

Constructors

Retomar1

Retomar2

Retomar3

RetomarM1

RetomarM2

RetomarM3

Sair1

Sair2

Sair3

SairM1

SairM2

SairM3

data Perdeu

Perdeu são as opções que aparecem quando o Jogador perde.

Constructors

Jogo1

Jogo2

Jogo3

Menu1

Menu2

Menu3

JogoM11

JogoM12

JogoM21

JogoM22

JogoM31

JogoM32

MenuM11

MenuM12

MenuM21

MenuM22

MenuM31

MenuM32

EmpataramJ1

EmpataramJ2

EmpataramJ3

EmpataramS1

EmpataramS2

EmpataramS3

data Modo

Modo são as opções que o Jogador tem, isto é, pode escolher jogar sozinho ou com mais um jogador.

Constructors

Single

Multi

Voltar

data Menu

Menu contem as diferentes opções de interação que o jogador pode ter para jogar.

Constructors

Opcoes Opcao

Perdeu Perdeu

Pausa Pausa

Modo Modo

Bot1

Bot2

Bot3

Inicio1

Inicio2

Inicio3

InicioM1

InicioM2

InicioM3

ModoJogo1

ModoJogo2

ModoJogo3

ModoJogoM1

ModoJogoM2

ModoJogoM3

Personagem1

Personagem2

Personagem3

PersonagemVoltar

PersonagemM1

PersonagemM2

PersonagemM3

PersonagemVoltarM

ModoInstrucoes1

ModoInstrucoesC

ModoInstrucoesS